Drastic changes to how RNG seeds are generated. They're always positive, and always between 0 and 2^(number of bits) - 1, where the number of bits is specified by the user (defaulting to 16).
git-svn-id: svn+ssh://battistini.cscs.lsa.umich.edu/users/gsweep/svn/gridsweeper@90 e466becf-6918-0410-8e4b-ca7f87a03805

+++ trunk/src/edu/umich/lsa/cscs/gridsweeper/Adapter.java
@@ -1,5 +1,7 @@
 package edu.umich.lsa.cscs.gridsweeper;
 
+import java.math.BigInteger;
+
 import edu.umich.lsa.cscs.gridsweeper.parameters.ParameterMap;
 
 /**
@@ -25,5 +27,5 @@ public interface Adapter
 	 * and standard error.
 	 * @throws AdapterException For any reason defined by the implementing class.
 	 */
-	public RunResults run(ParameterMap parameterMap, int runNumber, long rngSeed) throws AdapterException;
+	public RunResults run(ParameterMap parameterMap, int runNumber, BigInteger rngSeed) throws AdapterException;
 }
