+++ trunk/src/edu/umich/lsa/cscs/gridsweeper/ExperimentCase.java
@@ -1,5 +1,6 @@
 package edu.umich.lsa.cscs.gridsweeper;
 
+import java.math.BigInteger;
 import java.util.*;
 
 import edu.umich.lsa.cscs.gridsweeper.parameters.*;
@@ -13,14 +14,14 @@ import edu.umich.lsa.cscs.gridsweeper.parameters.*;
 public class ExperimentCase
 {
 	ParameterMap parameterMap;
-	List<Long> rngSeeds;
+	List<BigInteger> rngSeeds;
 	
 	/**
 	 * Initializes the experiment case with the parameter settings and random seeds.
 	 * @param parameterMap The parameter assignments.
 	 * @param rngSeeds The random seeds.
 	 */
-	public ExperimentCase(ParameterMap parameterMap, List<Long> rngSeeds)
+	public ExperimentCase(ParameterMap parameterMap, List<BigInteger> rngSeeds)
 	{
 		this.parameterMap = parameterMap;
 		this.rngSeeds = rngSeeds;
@@ -39,7 +40,7 @@ public class ExperimentCase
 	 * Getter for the list of random seeds.
 	 * @return The list of random seeds.
 	 */
-	public List<Long> getRngSeeds()
+	public List<BigInteger> getRngSeeds()
 	{
 		return rngSeeds;
 	}
