+++ trunk/src/edu/umich/lsa/cscs/gridsweeper/GridSweeperRunner.java
@@ -1,6 +1,7 @@
 package edu.umich.lsa.cscs.gridsweeper;
 
 import java.io.*;
+import java.math.BigInteger;
 
 import static edu.umich.lsa.cscs.gridsweeper.StringUtils.*;
 
@@ -67,7 +68,7 @@ public class GridSweeperRunner
 			// Run!
 			ParameterMap parameters = setup.getParameters();
 			int runNumber = setup.getRunNumber();
-			long rngSeed = setup.getRngSeed();
+			BigInteger rngSeed = setup.getRngSeed();
 			results = adapter.run(parameters, runNumber, rngSeed);
 			
 			// Stage files listed in run properties back to server (if asked for)
