+++ trunk/src/edu/umich/lsa/cscs/gridsweeper/RunSetup.java
@@ -1,6 +1,7 @@
 package edu.umich.lsa.cscs.gridsweeper;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.util.*;
 
 import edu.umich.lsa.cscs.gridsweeper.parameters.ParameterMap;
@@ -20,10 +21,10 @@ public class RunSetup implements Serializable
 	private String fileTransferSubpath;
 	private ParameterMap parameters;
 	private int runNumber;
-	private long rngSeed;
+	private BigInteger rngSeed;
 	private StringList outputFiles;
 	
-	public RunSetup(Settings settings, StringMap inputFiles, String fileTransferSubpath, ParameterMap parameters, int runNumber, long rngSeed, StringList outputFiles)
+	public RunSetup(Settings settings, StringMap inputFiles, String fileTransferSubpath, ParameterMap parameters, int runNumber, BigInteger rngSeed, StringList outputFiles)
 	{
 		this.settings = settings;
 		this.inputFiles = inputFiles;
@@ -39,7 +40,7 @@ public class RunSetup implements Serializable
 		return settings;
 	}
 	
-	public long getRngSeed()
+	public BigInteger getRngSeed()
 	{
 		return rngSeed;
 	}
